move output types to kebap-cased module
